pyproject.toml
[tool.aerich]
tortoise_orm = "infrastructure.database.connection.sample_conf"
location = "infrastructure/database/migrations"
src_folder = "./."

[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Dmitry Sidorenko", email = "dsidorenko@a-bt.ru"},
]
dependencies = [
    "aiofiles==0.8.0",
    "aiosqlite==0.17.0",
    "asyncpg==0.25.0",
    "click==8.1.2",
    "dictdiffer==0.9.0",
    "dnspython==2.2.1",
    "email-validator==1.1.3",
    "environs==9.5.0",
    "httptools==0.4.0",
    "idna==3.3",
    "iso8601==1.0.2",
    "loguru==0.6.0",
    "marshmallow==3.15.0",
    "multidict==6.0.2",
    "packaging==21.3",
    "pydantic==1.9.0",
    "pyparsing==3.0.8",
    "pypika-tortoise==0.1.4",
    "python-dotenv==0.20.0",
    "pytz==2022.1",
    "pyyaml==6.0",
    "sanic==22.3.0",
    "sanic-cors==2.0.1",
    "sanic-ext==22.3.0",
    "sanic-routing==22.3.0",
    "tomlkit==0.10.1",
    "tortoise-orm==0.19.0",
    "typing-extensions==4.1.1",
    "ujson==5.2.0",
    "uvloop==0.16.0",
    "websockets==10.2",
    "pycryptodome>=3.14.1",
    "aerich>=0.6.3",
    "aiosmtplib>=1.1.6",
    "celery[redis]>=5.2.6",
    "asgiref>=3.5.0",
    "flower>=1.0.0",
    "pytest-asyncio>=0.18.3",
    "sanic-testing>=22.3.0",
    "pyee>=9.0.4",
    "aioredis>=2.0.1",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
